requirement.txt
Django==3.1.1
djangorestframework-simplejwt==4.4.0
django-cors-headers==3.5.0
djangorestframework==3.11.1
djoser
numpy
Pillow==8.1.0
opencv-contrib-python
mediapipe==0.8.3.1
imutils==0.5.4